+++ airbyte-webapp/src/pages/SettingsPage/pages/ConfigurationsPage/index.tsx
@@ -1,0 +1,3 @@
+import ConfigurationsPage from "./ConfigurationsPage";
+
+export default ConfigurationsPage;
